now does passenger routing! uses a-star search on an implicit graph
partial fix for #5

No heuristic for now...

The implicit graph is determined from the Schedule & the TrackNetwork. Each
"vertex" is a pair of TrackNetwork::IDs and a time. Out edges are determined
from trains that leave in the future, and a corrisponding "vertex" is the
pair of the next ID the train goes to, and the time it gets there.

Still has some problems: the search algorithm needs to iterate all the out-edges
of a vertex, but this technically will be an infinite number. Need to come
up with a way to "hide" all except a small finite number.

code loosely based off of http://stackoverflow.com/a/9782649/2256231